test(tonk-concept): inline browser tests into render.rs mod tests
Match the rest of the workspace — wasm-bindgen tests live in
`#[cfg(test)] mod tests` inside the source module rather than a
separate `tests/` crate. Drops the standalone integration crate
and re-privatises `mod render`.
